refactor(app): move run() and newLogger() to internal/app (#32, step 3/3)
Closes the issue #32 refactor. cmd/topology-exporter/main.go is now a
21-line entry point that calls app.Run; everything else lives in
internal/app and internal/app/httpx.

main.go shrinkage across the three commits:
  1527 -> 1427 -> 297 -> 21 lines

New file:
  - internal/app/app.go - Run(), NewLogger()

Tests in main_test.go that exercised run()/newLogger() directly now call
app.Run / app.NewLogger; no test was deleted or skipped.

No behavior change. go build / go vet / go test -race ./... / gofmt all clean.
